Changing Errno::EAGAIN to IO::WaitReadable to fix failing build on Windows
This happens when the IO buffer is not ready and we try to read from it.
It throws Errno::EAGAIN or Errno::EWOULDBLOCK, on Linux and Mac these both errors are essentially the same so rescueing either of the two should be fine. However on widows these both are different and it causes the server to crash and the appveyor build to fail.
Ruby-doc says that both of these errors are extended by IO::WaitReadable, so instead of rescueing both these exceptions individually we can safely use IO::WaitReadable.